backend/requirements.txt
# GANPAN Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP Client
httpx==0.26.0

# Environment Variables
python-dotenv==1.0.0

# File Upload Support
python-multipart==0.0.6

# Async Support
aiofiles==23.2.1

# Image Processing (optional - for local preprocessing)
Pillow==10.2.0

# Utilities
pydantic==2.5.3
